+++ searchcore/src/tests/proton/common/timer/timer_test.cpp
@@ -1,12 +1,11 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include <vespa/vespalib/testkit/testapp.h>
-#include <vespa/vespalib/util/scheduledexecutor.h>
+#include <vespa/searchcore/proton/common/scheduledexecutor.h>
 #include <vespa/vespalib/util/size_literals.h>
 #include <vespa/fnet/transport.h>
 #include <vespa/fastos/thread.h>
 
-using namespace vespalib;
 using vespalib::Executor;
 typedef Executor::Task Task;
 
@@ -28,7 +27,7 @@ TEST("testScheduling") {
     FastOS_ThreadPool threadPool(64_Ki);
     FNET_Transport transport;
     transport.Start(&threadPool);
-    ScheduledExecutor timer(transport);
+    proton::ScheduledExecutor timer(transport);
     timer.scheduleAtFixedRate(std::make_unique<TestTask>(latch1), 100ms, 200ms);
     timer.scheduleAtFixedRate(std::make_unique<TestTask>(latch2), 500ms, 500ms);
     EXPECT_TRUE(latch1.await(60s));
@@ -42,7 +41,7 @@ TEST("testReset") {
     FastOS_ThreadPool threadPool(64_Ki);
     FNET_Transport transport;
     transport.Start(&threadPool);
-    ScheduledExecutor timer(transport);
+    proton::ScheduledExecutor timer(transport);
     timer.scheduleAtFixedRate(std::make_unique<TestTask>(latch1), 2s, 3s);
     timer.reset();
     EXPECT_TRUE(!latch1.await(3s));
