+++ searchcore/src/vespa/searchcore/proton/common/CMakeLists.txt
@@ -25,6 +25,7 @@ vespa_add_library(searchcore_pcommon STATIC
     selectpruner.cpp
     state_reporter_utils.cpp
     statusreport.cpp
+    scheduledexecutor.cpp
     DEPENDS
     searchcore_proton_metrics
     searchcore_fconfig
